+++ edge/sync.ts
@@ -1,9 +1,11 @@
+import * as _ from 'lodash'
 import * as vscode from 'vscode'
 
 import * as Util from './Util'
 import * as Git from './Git'
+import { getMergedBranchNames } from './deleteMergedBranches'
 
-export default async function () {
+export default async function (options: { token: vscode.CancellationToken }) {
 	const workspace = await Util.getCurrentWorkspace()
 	if (!workspace) {
 		return null
@@ -23,16 +25,39 @@ export default async function () {
 		}
 	}
 
-	await vscode.window.withProgress({ location: vscode.ProgressLocation.Window, title: 'Syncing...' }, async () => {
+	await vscode.window.withProgress({ location: vscode.ProgressLocation.Window, title: 'Syncing...' }, async ({ report }) => {
 		try {
 			if (status.remote === '') {
-				await Git.run(workspace.uri, 'push', 'origin', status.local)
+				await Git.run(workspace.uri, 'push', 'origin', status.local, { retry: 1 })
 				await Git.setRemoteBranch(workspace.uri, status.local)
 			}
 
+			await Git.run(workspace.uri, 'push', '--tags', { retry: 1 })
+
 			await Git.run(workspace.uri, 'pull', '--all', '--rebase')
-			await Git.run(workspace.uri, 'push', '--all')
-			await Git.run(workspace.uri, 'push', '--tags')
+
+			const counterparts = await Git.getBranchCounterparts(workspace.uri)
+			const remoteBranches = new Set(await Git.getRemoteBranchNames(workspace.uri))
+			const mergedBranches = new Set(await getMergedBranchNames(workspace.uri, false))
+			for (const { local, remote } of counterparts) {
+				if (local === status.local) {
+					continue
+				}
+
+				if (remoteBranches.has(remote)) {
+					const groups = await Git.getBranchTopology(workspace.uri, local, remote)
+					if (groups.length === 2) {
+						await Git.run(workspace.uri, 'push', 'origin', local, { retry: 1 })
+					}
+
+				} else if (mergedBranches.has(local)) {
+					await Git.run(workspace.uri, 'branch', '--delete', '--force', local, { retry: 1 })
+
+				} else {
+					await Git.run(workspace.uri, 'push', 'origin', local, { retry: 1 })
+					await Git.setRemoteBranch(workspace.uri, local)
+				}
+			}
 
 		} catch (ex) {
 			Util.setWorkspaceAsFirstTryNextTime(workspace)